Cleanups and integration of DataDictParallel with PipelineManager
