+++ ptp/utils/app_state.py
@@ -70,7 +70,7 @@ class AppState(metaclass=SingletonMetaClass):
         self.set_cpu_types()
         self.use_gpu = False
         self.use_dataparallel = False
-        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+        self.device = torch.device('cpu')
 
         # Reset global counters.
         self.epoch = None # Processor is not using the notion of epoch.
@@ -86,9 +86,11 @@ class AppState(metaclass=SingletonMetaClass):
         if torch.cuda.is_available() and self.args.use_gpu:
             self.logger.info('Running computations on GPU using CUDA')
             self.set_gpu_types()
-            #self.device = torch.device('cuda:0')
+            # Use GPU.
             self.use_gpu = True
-            if self.args.use_dataparallel:
+            self.device = torch.device('cuda')
+            # Use DataParallel if more than 1 device is available.
+            if torch.cuda.device_count() > 1:
                 self.use_dataparallel = True
         elif self.args.use_gpu:
             self.logger.warning('GPU utilization is demanded but there are no available GPU devices! Using CPUs instead')
